Create module-1.cs
For items of work that were awarded in any of the preceding 3 years, estimate shall be
calculated as the average of such awarded rates. However, in case the awarded rate of any
previous year under consideration was less than -10% or more than +10%of the estimated rate,
the reduction below -10% or excess over of +10% shall be ignored in the summation of the
awarded rates (for example 12% will be taken as -10% and +12% will be taken as +10% for the
purpose of average). Also if an item of work was awarded in any 1 or 2 of the preceding 3 years,
average awarded rates for those years will be considered for estimate calculation. While
averaging, the labour escalation as applicable in the latest awarded rate/contract shall be added
to the previously awarded rates (to the extent not added earlier) to bring them to the same labour
rate level. The same shall be verified by Estimates Cell.
In a particular department, if an item was awarded in more than one contract in a year, the
average of the awarded rates for that particular year(after ignoring any excess over +10% or
reduction below -10%) shall be considered as the awarded rate for that item in that year.
Estimate shall be calculated as the average of such awarded rates of the preceding 3 years (to
the extent available). While averaging, the labor escalation as applicable in the latest awarded
rate/contract shall be added to the previously awarded rates (to the extent not added earlier) to
bring them to the same labor rate level. The same shall be verified by Estimates Cell.
In case of a long term contract covering previous 3 years, the awarded rate with latest eligible
labor escalation (to the extent not added earlier) shall be considered as estimate for the present.
Further, in case more than one long term contract and/or annual contract fall within previous 3
years, then average of the awarded rates with latest eligible labour escalation (to the extent not
added earlier) shall be considered as estimate for the present. Any excess over +10% or
reduction below -10% shall be ignored in the process. The same shall be verified by Estimates
Cell.
In case of considering the extra items for which the rates were previously approved, as BOQ
items in any subsequent tender proposal, the department shall add the applicable escalations
to those rates also. Departments shall furnish detailed calculation for escalation along with the
copies of the approvals for such extra items to enable Estimates Cell to check the escalation
calculation.